Fix poly1305-x86.pl
Imported from patch attached to
https://rt.openssl.org/Ticket/Display.html?id=4439

The root problem appears to be that lazy_reduction tries to use paddd
instead of paddq when they believe the sum will not overflow a u32. In
the final call to lazy_reduction, this is not true. svaldez and davidben
attempted to work through the bounds, but the bounds derived from the cited
paper imply paddd is always fine. Empirically in a debugger, the bounds are exceeded in the test
case.

davidben requested more comments from upstream on the bug. When upstream lands
their final fix (hopefully with comments), I will update this code. In the
meantime,
let's stop carrying known-broken stuff.

(vlazy_reduction is probably something similar, but since we don't
enable that
code, we haven't bothered analyzing it.)

From BoringSSL