Avoid an overflow in constructing the ServerKeyExchange message
We calculate the size required for the ServerKeyExchange message and
then call BUF_MEM_grow_clean() on the buffer. However we fail to take
account of 2 bytes required for the signature algorithm and 2 bytes for
the signature length, i.e. we could overflow by 4 bytes. In reality this
won't happen because the buffer is pre-allocated to a large size that
means it should be big enough anyway.

From OpenSSL