Fix DTLS unprocessed records bug
During a DTLS handshake we may get records destined for the next epoch
arrive before we have processed the CCS. In that case we can't decrypt
or
verify the record yet, so we buffer it for later use. When we do receive
the CCS we work through the queue of unprocessed records and process
them.

Unfortunately the act of processing wipes out any existing packet data
that we were still working through. This includes any records from the
new
epoch that were in the same packet as the CCS. We should only process
the
buffered records if we've not got any data left.

From OpenSSL